+++ Sources/EventsTracker/EventsCache.swift
@@ -83,7 +83,7 @@ class EventsCache<T: CacheableEvent> {
         
         if writeToDiskTimer == nil {
             writeToDiskTimer = Timer(timeInterval: 0.2, target: self, selector: #selector(writeToDiskTimerTick(_:)), userInfo: nil, repeats: false)
-            RunLoop.main.add(writeToDiskTimer!, forMode: RunLoopMode.commonModes)
+            RunLoop.main.add(writeToDiskTimer!, forMode: RunLoop.Mode.common)
         }
     }
     @objc fileprivate func writeToDiskTimerTick(_ timer: Timer) { writeCurrentStateToDisk() }