+++ Sources/EventsTracker/EventsTracker+SessionLifecycleHandler.swift
@@ -8,6 +8,7 @@
 //  Copyright (c) 2018 ShopGun. All rights reserved.
 
 import Foundation
+import UIKit
 
 extension EventsTracker {
     
@@ -17,12 +18,12 @@ extension EventsTracker {
         var didStartNewSession: (() -> Void)?
         
         init() {
-            NotificationCenter.default.addObserver(self, selector: #selector(appDidBecomeActive(_:)), name: .UIApplicationDidBecomeActive, object: nil)
-            NotificationCenter.default.addObserver(self, selector: #selector(appDidEnterBackground(_:)), name: .UIApplicationDidEnterBackground, object: nil)
+            NotificationCenter.default.addObserver(self, selector: #selector(appDidBecomeActive(_:)), name: UIApplication.didBecomeActiveNotification, object: nil)
+            NotificationCenter.default.addObserver(self, selector: #selector(appDidEnterBackground(_:)), name: UIApplication.didEnterBackgroundNotification, object: nil)
         }
         deinit {
-            NotificationCenter.default.removeObserver(self, name: .UIApplicationDidBecomeActive, object: nil)
-            NotificationCenter.default.removeObserver(self, name: .UIApplicationDidEnterBackground, object: nil)
+            NotificationCenter.default.removeObserver(self, name: UIApplication.didBecomeActiveNotification, object: nil)
+            NotificationCenter.default.removeObserver(self, name: UIApplication.didEnterBackgroundNotification, object: nil)
         }
         
         fileprivate var isInBackground: Bool = false